requirements.txt
torch
pytorch_lightning
matplotlib
numpy
trimesh
hydra-core
omegaconf
tqdm
pandas
invoke
rtree
scipy
torch-geometric
wandb
e3nn